+++ rust/operator-binary/src/catalog/iceberg.rs
@@ -1,12 +1,12 @@
 use super::{config::CatalogConfig, ExtendCatalogConfig, FromTrinoCatalogError, ToCatalogConfig};
 use async_trait::async_trait;
 use stackable_operator::client::Client;
-use stackable_trino_crd::catalog::hive::HiveConnector;
+use stackable_trino_crd::catalog::iceberg::IcebergConnector;
 
-pub const CONNECTOR_NAME: &str = "hive";
+pub const CONNECTOR_NAME: &str = "iceberg";
 
 #[async_trait]
-impl ToCatalogConfig for HiveConnector {
+impl ToCatalogConfig for IcebergConnector {
     async fn to_catalog_config(
         &self,
         catalog_name: &str,
@@ -16,11 +16,11 @@ impl ToCatalogConfig for HiveConnector {
         let mut config = CatalogConfig::new(catalog_name.to_string(), CONNECTOR_NAME);
 
         // No authorization checks are enforced at the catalog level.
-        // We don't want the hive connector to prevent users from dropping tables.
-        // We also don't want that the hive connector makes decisions on which user is allowed to do what.
+        // We don't want the iceberg connector to prevent users from dropping tables.
+        // We also don't want that the iceberg connector makes decisions on which user is allowed to do what.
         // This decision should be done globally (for all catalogs) by OPA.
-        // See https://trino.io/docs/current/connector/hive-security.html
-        config.add_property("hive.security", "allow-all");
+        // See https://trino.io/docs/current/connector/iceberg.html
+        config.add_property("iceberg.security", "allow-all");
 
         self.metastore
             .extend_catalog_config(&mut config, catalog_name, catalog_namespace.clone(), client)